% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glyph_bitmap.R
\name{glyph_bitmap}
\alias{glyph_bitmap}
\title{Extracting bitmap glyph from fonts loaded by the 'sysfonts' package}
\usage{
glyph_bitmap(ch = "A", family = "mono", face = "regular",
  pixel_size = 50, rot = 0)
}
\arguments{
\item{ch}{A character string whose glyph is to be extracted.}

\item{family}{The family of the font. See
\code{\link[sysfonts]{font.families}()}.}

\item{face}{The font face. Possible values are "regular", "bold", "italic"
and "bolditalic".}

\item{pixel_size}{The pixel size of the font.}

\item{rot}{The angle in degrees of the counterclockwise string rotation.}
}
\value{
A matrix with class "glyph_bitmap" that represents the bitmap glyph.
        Values in the matrix are integers ranging from 0 to 255. Larger
        number indicates a darker color.
}
\description{
This function returns a bitmap matrix that represents the glyph of a given
string with the specified font.
}
\examples{
f = glyph_bitmap("fontr", family = "serif", face = "italic")
plot(f)
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

