% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glyph_outline.R
\name{glyph_outline}
\alias{glyph_outline}
\title{Extracting the glyph outline of a character}
\usage{
glyph_outline(ch = "a", family = "sans", face = "regular")
}
\arguments{
\item{ch}{A character whose glyph is to be extracted.}

\item{family}{The family of the font. See
\code{\link[sysfonts]{font.families}()}.}

\item{face}{The font face. Possible values are "regular", "bold", "italic"
              and "bolditalic".}
}
\value{
A data frame with class "glyph_outline".
        Each row of the data frame contains two numeric values, giving the
        x and y coordinates of a point, and a character, indicating the action
        to take on that point in order to draw the outline.
        The meaning of different characters are given below:

        \item{M}{Move the pen to the point}
        \item{L}{Draw a straight line to the point}
        \item{Q}{Draw a quadratic Bézier curve to the point, followed by a
                 "B" row that gives the coordinates of the control point}
        \item{C}{Draw a cubic Bézier curve to the point, followed by two
                 "B" rows that give the coordinates of the control points}

        More details can be found in sections 8.3.4 to 8.3.7 of the SVG
        specification. (\url{http://www.w3.org/TR/SVG/paths.html})
}
\description{
This function extracts the outline of character glyphs contained in a font.
}
\examples{
R = glyph_outline("R", family = "sans", face = "bold")
head(R, 20)
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

