% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glyph_outline.R
\name{glyph_polygon}
\alias{glyph_polygon}
\title{Extracting the glyph outline of a character as polygons}
\usage{
glyph_polygon(ch = "a", family = "sans", face = "regular", nseg = 10)
}
\arguments{
\item{ch}{A character whose glyph is to be extracted.}

\item{family}{The family of the font. See
\code{\link[sysfonts]{font.families}()}.}

\item{face}{The font face. Possible values are "regular", "bold", "italic"
and "bolditalic".}

\item{nseg}{The number of line segments to approximate
              a piece of curve in the outline. The larger \code{nseg} is,
              the smoother text outlines would be. Usually a value
              between 5~20 would be enough.}
}
\value{
A data frame with class "glyph_polygon" that contains two columns,
        giving the x and y coordinates of the outline. Rows that contain
        \code{NA} represents the start of a new polygon path.
}
\description{
This function extracts the outline of character glyphs contained in a font,
and converts the outline into polygons.
}
\examples{
R = glyph_polygon("R", family = "sans", face = "bold")
plot(R)
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

