% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glyph_outline.R
\name{plot.glyph_polygon}
\alias{plot.glyph_polygon}
\title{Plotting function for glyph outline in polygon form}
\usage{
\method{plot}{glyph_polygon}(glyph, ...)
}
\description{
This function plots glyph outline polygons of class "glyph_polygon", typically returned
by the function \code{\link{glyph_polygon}()}.
}
\examples{
R = glyph_polygon("R", family = "sans", face = "bold")
plot(R)
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
\seealso{
\code{\link{glyph_polygon}()}
}

