% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glyph_bitmap.R
\name{plot.glyph_bitmap}
\alias{plot.glyph_bitmap}
\title{Plotting function for bitmap glyph}
\usage{
\method{plot}{glyph_bitmap}(glyph, ...)
}
\description{
This function plots a bitmap glyph of class "glyph_bitmap", typically returned
by the function \code{\link{glyph_bitmap}()}.
}
\examples{
f = glyph_bitmap("fontr", family = "serif", face = "italic")
plot(f)
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
\seealso{
\code{\link{glyph_bitmap}()}
}

